splitting tweets about votes and contributions into separate functions
